[refactor] Restructure and clean UsersListViewController
