+++ miniPostApp/miniPost/Modules/Users/UIKit/UsersListViewController.swift
@@ -11,8 +11,7 @@ protocol UIUsersListView: AnyObject {
     func didLoadData()
 }
 
-class UsersListViewController: UIViewController, UICollectionViewDelegateFlowLayout, UICollectionViewDataSource, UICollectionViewDelegate {
-    
+class UsersListViewController: UIViewController {
     private let collectionView: UICollectionView = {
         let cv = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
         cv.register(UsersCollectionCell.self, forCellWithReuseIdentifier: UsersCollectionCell.identifier)
@@ -31,12 +30,12 @@ class UsersListViewController: UIViewController, UICollectionViewDelegateFlowLay
     
     private let configurator: UsersConfigurator = .init()
     var presenter: UsersPresenter?
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         configurator.configure(viewController: self)
-
+        
         view.addSubview(collectionView)
         NSLayoutConstraint.activate([
             view.leadingAnchor.constraint(equalTo: collectionView.leadingAnchor),
@@ -46,7 +45,7 @@ class UsersListViewController: UIViewController, UICollectionViewDelegateFlowLay
         ])
         collectionView.delegate = self
         collectionView.dataSource = self
-
+        
         view.addSubview(loadingView)
         view.addConstraint(loadingView.topAnchor.constraint(equalTo: view.topAnchor))
         
@@ -54,23 +53,29 @@ class UsersListViewController: UIViewController, UICollectionViewDelegateFlowLay
             await presenter?.viewDidLoad()
         }
     }
+}
+
+extension UsersListViewController: UIUsersListView {
+    func didLoadData() {
+        onDataLoaded()
+    }
     
     @MainActor
     private func onDataLoaded() {
         collectionView.reloadData()
         loadingView.isHidden = true
     }
+}
 
-    // MARK: - UICollectionViewDataSource
-    
+extension UsersListViewController: UICollectionViewDataSource {
     func numberOfSections(in collectionView: UICollectionView) -> Int {
         return 1
     }
-
+    
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return presenter?.users.count ?? 0
     }
-
+    
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         guard let presenter,
               indexPath.row < presenter.users.count,
@@ -85,74 +90,24 @@ class UsersListViewController: UIViewController, UICollectionViewDelegateFlowLay
         
         return cell
     }
-    
-    // MARK: - UICollectionViewDelegate
-    
+}
+
+extension UsersListViewController: UICollectionViewDelegate {
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         guard let presenter, indexPath.row < presenter.users.count else { return }
         
         showUserModal(user: presenter.users[indexPath.row])
     }
-
+    
     private func showUserModal(user: User) {
         view.addSubview(UserView(frame: view.frame, userName: user.name, userPhone: user.phone))
     }
+}
 
-    // MARK: - UICollectionViewDelegateFlowLayout
-
+extension UsersListViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         let width = collectionView.frame.width
         let height: CGFloat = 90
         return CGSize(width: width, height: height)
-    }
-}
-
-class UserView: UIView {
-    init(frame: CGRect, userName: String, userPhone: String) {
-        super.init(frame: frame)
-        self.backgroundColor = .white
-
-        let userNameLabel = UILabel()
-        userNameLabel.text = userName
-        self.addSubview(userNameLabel)
-        userNameLabel.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            userNameLabel.centerXAnchor.constraint(equalTo: self.centerXAnchor),
-            userNameLabel.centerYAnchor.constraint(equalTo: self.centerYAnchor, constant: -20)
-        ])
-
-        let userPhoneLabel = UILabel()
-        userPhoneLabel.text = userPhone
-        self.addSubview(userPhoneLabel)
-        userPhoneLabel.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            userPhoneLabel.centerXAnchor.constraint(equalTo: self.centerXAnchor),
-            userPhoneLabel.centerYAnchor.constraint(equalTo: self.centerYAnchor)
-        ])
-        
-        let closeButton = UIButton(type: .custom)
-        closeButton.setTitle("Close", for: .normal)
-        closeButton.setTitleColor(.blue, for: .normal)
-        closeButton.addTarget(self, action: #selector(closeButtonTapped), for: .touchUpInside)
-        self.addSubview(closeButton)
-        closeButton.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            closeButton.centerXAnchor.constraint(equalTo: self.centerXAnchor),
-            closeButton.centerYAnchor.constraint(equalTo: self.centerYAnchor, constant: 20)
-        ])
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    @objc func closeButtonTapped() {
-        self.removeFromSuperview()
-    }
-}
-
-extension UsersListViewController: UIUsersListView {
-    func didLoadData() {
-        onDataLoaded()
     }
 }
